+++ packages/sandbox/src/index.ts
@@ -115,16 +115,16 @@ export const createTypeScriptSandbox = (
   monaco: Monaco,
   ts: typeof import("typescript")
 ) => {
-  const config = { ...defaultPlaygroundSettings(), ...partialConfig }
-  if (!("domID" in config) && !("elementToAppend" in config))
+  if (!("domID" in partialConfig) && !("elementToAppend" in partialConfig))
     throw new Error("You did not provide a domID or elementToAppend")
 
+  const config = { ...defaultPlaygroundSettings(), ...partialConfig }
   const defaultText = config.suppressAutomaticallyGettingDefaultText
     ? config.text
     : getInitialCode(config.text, document.location)
 
   // Defaults
-  const compilerDefaults = getDefaultSandboxCompilerOptions(config, monaco)
+  const compilerDefaults = getDefaultSandboxCompilerOptions(config, monaco, ts)
 
   // Grab the compiler flags via the query params
   let compilerOptions: CompilerOptions
@@ -146,7 +146,10 @@ export const createTypeScriptSandbox = (
 
   const language = languageType(config)
   const filePath = createFileUri(config, compilerOptions, monaco)
-  const element = "domID" in config ? document.getElementById(config.domID) : (config as any).elementToAppend
+  const element = "elementToAppend" in config ? config.elementToAppend : document.getElementById(config.domID)
+  if (!element)
+    throw new Error("DOM element lookup by domID failed")
+
 
   const model = monaco.editor.createModel(defaultText, language, filePath)
   monaco.editor.defineTheme("sandbox", sandboxTheme)